New arguments format, new controller, deprecation warning
Change the format of the arguments. Use duration instead of
integers to allow setting arbitrary cleanup periods. #44

Rename controller from PodController into Kleaner for clear
indication that it operates not only on pods, but also on jobs
(and whatever will be introduced later).

Kleaner now watches not only for pods, but also for jobs, and removes
both. #43 #38