Separate pod and job decision logic, add tests
Introduce `ignore-owned-by-cronjobs` flag to explicitly opt-in for the change of
logic introduced in 7e342c16f94d6e8aec1a0a7811b210b1bcf14e6b. Flag is marked
as `EXPERIMENTAL`, which means it could be changed or renamed in the coming versions.

Split pod and job decision logic to be able to unit test it. Add tests.

Signed-off-by: Sergey Nuzhdin <ipaq.lw@gmail.com>